WDID : - Add error handling for delete, archive/dis-archive, getProducts, conversions - Appels aux uses cases - Comments for - ProductsViewModel
WTD :
- String en Double ou Int
- Create an undo delete
- Add category variable to Product
- Configure budget
- Configure sorting buttons

+++ app/src/main/java/com/projects/writeit/feature_product/presentation/list_product/ProductsViewModel.kt
@@ -19,6 +19,7 @@ import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asSharedFlow
+import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.map
@@ -43,24 +44,24 @@ class ProductsViewModel @Inject constructor(
     private val productUseCases: ProductUseCases
 ) : ViewModel() {
 
+    // Etat des éléments faisant parti de l'écran de gestion des produits.
     private val _state = mutableStateOf(ProductsState())
     val state: State<ProductsState> = _state
 
+    // Produit récemment supprimé à ré-intégrer dans une méthode "UNDO" -> TODO ()
     private var recentlyDeletedProduct: Product? = null
 
+    // Travail asynchrone qui permet de contrôler la récupération de la liste des produits (actifs/archivés).
     private var getActiveProductsJob: Job? = null
     private var getArchivedProductsJob: Job? = null
 
+    // Flow permettant d'émettre des évènements uniques vers l'UI.
     private val _eventFlow = MutableSharedFlow<UiEvent>()
     val eventFlow = _eventFlow.asSharedFlow()
 
-    // Produit dont l'id est null et qio sert de réceptacle pour le produit à éditer.
+    // Produit dont l'id est null et qui sert de réceptacle pour le produit à éditer.
     private val _productToEdit = mutableStateOf<Product?>(null)
     val productToEdit: State<Product?> = _productToEdit
-
-    fun productToEdit(product: Product) {
-        _productToEdit.value = product
-    }
 
     // Au lancement du ViewModel, on récupère les produits actifs et archivés.
     init {
@@ -87,6 +88,11 @@ class ProductsViewModel @Inject constructor(
         }
         .stateIn(viewModelScope, SharingStarted.Eagerly, 0.0)
 
+    // On récupère les informations du produit à éditer.
+    fun productToEdit(product: Product) {
+        _productToEdit.value = product
+    }
+
 
     // Récupère les produits actifs depuis la couche UseCases,
     // puis les transforme en liste de SelectableProduct pour l'UI.
@@ -100,6 +106,8 @@ class ProductsViewModel @Inject constructor(
                 selectableActiveProducts = selectableList,
                 productsOrder = productOrder
             )
+        }.catch {
+            _eventFlow.emit(UiEvent.ShowSnackBar("La liste n'a pas pu être récupérée, patientez un instant..."))
         }.launchIn(viewModelScope)
         Log.i("DEBUG", "Produits: ${state.value.selectableActiveProducts.map { it.product.name }}")
     }
@@ -112,6 +120,8 @@ class ProductsViewModel @Inject constructor(
                 archivedProducts = it,
                 productsOrder = productOrder
             )
+        }.catch {
+            _eventFlow.emit(UiEvent.ShowSnackBar("La liste n'a pas pu être récupérée, patientez un instant..."))
         }.launchIn(viewModelScope)
         Log.i(
             "ARCHIVE DEBUG",
@@ -132,6 +142,7 @@ class ProductsViewModel @Inject constructor(
     //------------------------------------
     fun onEvent(event: ProductsEvent) {
         when (event) {
+            // Appelle la liste de produits de manière ordonnée.
             is ProductsEvent.Order -> {
                 if (state.value.productsOrder::class == event.productOrder::class &&
                     state.value.productsOrder.orderType == event.productOrder.orderType
@@ -141,37 +152,53 @@ class ProductsViewModel @Inject constructor(
                 getActiveProducts(event.productOrder)
             }
 
+            // Active/Désactive le dialogue d'édition de produit.
             is ProductsEvent.ToggleBottomDialog -> {
                 _state.value = state.value.copy(
                     showBottomSheet = !state.value.showBottomSheet
                 )
             }
 
+            // Active/Désactive le menu déroulant de tri.
             is ProductsEvent.ToggleSortDropDownMenu -> {
                 _state.value = state.value.copy(
                     sortDropDownExpanded = !state.value.sortDropDownExpanded
                 )
             }
 
+            // Archive le produit en lui attribuant la propriété "isArchived"
+            // On insère le produit dans la liste des produits archivés.
+            // Affiche un snackbar pour confirmer l'insertion / notifier l'échec de l'insertion.
             is ProductsEvent.ArchiveProduct -> {
+                val archivedProduct = event.product.copy(isArchived = true)
                 viewModelScope.launch {
-                    val archivedProduct = event.product.copy(isArchived = true)
-                    productUseCases.insertProduct(archivedProduct)
-                    _eventFlow.emit(UiEvent.ShowSnackBar("${archivedProduct.name} a été archivé"))
+                    try {
+                        productUseCases.insertProduct(archivedProduct)
+                        _eventFlow.emit(UiEvent.ShowSnackBar("${archivedProduct.name} a été archivé"))
+                    } catch (e : Exception) {
+                        _eventFlow.emit(UiEvent.ShowSnackBar("${archivedProduct.name} n'a pas été archivé"))
+                    }
                 }
             }
 
+            // Onb ré-insère le produit en lui otant la propriété "isArchived"
+            // On le remet à nouveau dans la liste des produits actifs.
+            // Affiche un snackbar pour confirmer l'insertion / notifier l'échec de l'insertion.
             is ProductsEvent.DisArchiveProduct -> {
+                val disArchiveProduct = event.product.copy(isArchived = false)
                 viewModelScope.launch {
-                    val disArchiveProduct = event.product.copy(isArchived = false)
-                    productUseCases.insertProduct(disArchiveProduct)
-                    _eventFlow.emit(UiEvent.ShowSnackBar("${disArchiveProduct.name} est de nouveau dans ta liste"))
+                    try {
+                        productUseCases.insertProduct(disArchiveProduct)
+                        _eventFlow.emit(UiEvent.ShowSnackBar("${disArchiveProduct.name} est de nouveau dans ta liste"))
+                    } catch (e : Exception) {
+                        _eventFlow.emit(UiEvent.ShowSnackBar("${disArchiveProduct.name} n'a pas pu être réintégré"))
+                    }
                 }
             }
 
             // Restaure le dernier produit supprimé (stocké temporairement dans `recentlyDeletedProduct`).
             // Réinsère le produit dans la base et réinitialise le cache local.
-            // Affiche une snackbar pour confirmer la restauration.
+            // Affiche un snackbar pour confirmer la restauration.
             is ProductsEvent.RestoreProduct -> {
                 viewModelScope.launch {
                     productUseCases.insertProduct(recentlyDeletedProduct ?: return@launch)
@@ -182,7 +209,7 @@ class ProductsViewModel @Inject constructor(
 
             // Restaure tous les produits archivés en les marquant comme non archivés.
             // Pour chaque produit, on l’insère de nouveau dans la base de données avec `isArchived = false`.
-            // Affiche une snackbar globale une fois l'opération terminée.
+            // Affiche un snackbar globale une fois l'opération terminée.
             is ProductsEvent.RestoreAllProducts -> {
                 viewModelScope.launch {
                     productUseCases.getArchivedProducts()
@@ -230,25 +257,51 @@ class ProductsViewModel @Inject constructor(
             // puis met à jour la liste et désactive le mode de sélection.
             is ProductsEvent.DeleteSelectedProducts -> {
                 viewModelScope.launch {
-                    val selectedItemsList = state.value.selectableActiveProducts
+                    var undeletedProducts = 0
+                    val itemsToDelete = state.value.selectableActiveProducts
                         .filter { it.isChecked }.map {
                             it.product
                         }
+                    // -> Si aucun item n'est sélectionné, on le notifie à l'utilisateur
+                    if (itemsToDelete.isEmpty()) {
+                        _eventFlow.emit(UiEvent.ShowSnackBar("Sélectionnez au moins un produit à supprimer"))
+                    } else {
+                        var updatedList : List<SelectableProduct> = emptyList()
 
-                    selectedItemsList.forEach {
-                        productUseCases.deleteProduct(it)
+                        itemsToDelete.forEach {
+                            try {
+                                productUseCases.deleteProduct(it)
+                            } catch (e: Exception) {
+                                undeletedProducts++
+                            }
+                        }
+                        // Si tous les produits ont été supprimés on désactive le mode suppression
+                        // et on le notifie à l'utilisateur.
+                        if(undeletedProducts == 0){
+                            updatedList = state.value.selectableActiveProducts
+                                .filter {!it.isChecked }
+                                .map { it.copy(isChecked = false)}
+                            _state.value = state.value.copy(
+                                selectableActiveProducts = updatedList,
+                                isSelectionMode = false,
+                                buttonDeleteIsVisible = false
+                            )
+                            _eventFlow.emit(UiEvent.ShowSnackBar("${itemsToDelete.size - undeletedProducts} produit(s) supprimé(s)"))
+                        } else {
+                            // Liste des produits non sélectionnés pour la suppression
+                            // qu'on souhaite récupérer pour une nouvelle suppression
+                            val failedProducts = state.value.selectableActiveProducts
+                                .filter {it.isChecked}
+                                .map {it.product}
+                            _state.value = state.value.copy(
+                                selectableActiveProducts = updatedList,
+                                isSelectionMode = failedProducts.isNotEmpty(),  // Reste en mode selection si erreurs
+                                buttonDeleteIsVisible = failedProducts.isNotEmpty()
+                            )
+                            _eventFlow.emit(UiEvent.ShowSnackBar("${itemsToDelete.size - undeletedProducts} produit(s) supprimé(s), $undeletedProducts en échec"))
+                        }
+
                     }
-                    Log.d("SNACKBAR", "Suppression détectée, affichage snackbar")
-                    _eventFlow.emit(UiEvent.ShowSnackBar("${selectedItemsList.size} produit(s) supprimé(s)"))
-                    val updatedList = state.value.selectableActiveProducts
-                        .filter { !it.isChecked }
-                        .map { it.copy(isChecked = false) }
-                    _state.value = state.value.copy(
-                        selectableActiveProducts = updatedList,
-                        isSelectionMode = false,
-                        buttonDeleteIsVisible = false
-                    )
-
                 }
             }
         }
